lab7/pyproject.toml
[tool.poetry]
name = "sometrainindrf"
version = "0.1.0"
description = ""
authors = ["syrlybajramazan <syrlybai056@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
djangorestframework = "^3.14.0"
pillow = "^10.2.0"
django-environ = "^0.11.2"
requests = "^2.31.0"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
